refactor(www): refactor folder structure
