+++ apps/www/src/app/[locale]/(blog)/notes/page.tsx
@@ -4,10 +4,9 @@ import Link from "next/link";
 import Image from "@chia/ui/image";
 import ImageZoom from "@chia/ui/image-zoom";
 
+import { List } from "@/components/blog/notes";
 import { getNotes } from "@/services/feeds.service";
 import { PageParamsWithLocale } from "@/utils/i18n";
-
-import { List } from "../_components/notes";
 
 export const metadata: Metadata = {
   title: "Notes",
