+++ apps/www/src/app/[locale]/@modal/(.)email/page.tsx
@@ -1,5 +1,3 @@
-import "@/app/[locale]/contact/style.css";
-
 import Form from "./form";
 
 const Page = () => {
