+++ apps/www/src/components/commons/background.tsx
