+++ legacy/service/src/modules/auth/auth.controller.ts
@@ -7,7 +7,7 @@ import {
   HttpException,
   Query,
   Param,
-} from "@nestjs/common";
+} from "@nestjs/commons";
 import { ApiOperation, ApiResponse, ApiTags } from "@nestjs/swagger";
 import {
   Request as ExpressRequest,
