+++ legacy/service/src/modules/auth/auth.module.ts
@@ -1,4 +1,4 @@
-import { Module } from "@nestjs/common";
+import { Module } from "@nestjs/commons";
 
 import AuthController from "./auth.controller";
 import AuthService from "./auth.service";
