+++ legacy/service/src/modules/auth/auth.service.ts
@@ -2,7 +2,7 @@ import {
   Injectable,
   InternalServerErrorException,
   HttpException,
-} from "@nestjs/common";
+} from "@nestjs/commons";
 import {
   Request as ExpressRequest,
   Response as ExpressResponse,
