+++ legacy/service/src/modules/feed/feed.controller.ts
@@ -5,7 +5,7 @@ import {
   InternalServerErrorException,
   HttpException,
   Query,
-} from "@nestjs/common";
+} from "@nestjs/commons";
 import { ApiOperation, ApiResponse, ApiTags } from "@nestjs/swagger";
 import { isArray } from "class-validator";
