+++ legacy/service/src/modules/rate-limiter/rate-limiter.module.ts
@@ -1,4 +1,4 @@
-import { Module } from "@nestjs/common";
+import { Module } from "@nestjs/commons";
 
 import RateLimiterController from "./rate-limiter.controller";
